Changed beta from an integer to a float

+++ hawk_eye/train/calculate_f1_score.py
@@ -1,5 +1,5 @@
 def calculate_f1_score(
-    true_positives, false_positives, false_negatives, beta: int
+    true_positives, false_positives, false_negatives, beta: float
 ) -> int:
 
     """ This function finds the F1 score for a model.
